Fix stash merge conflicts

+++ app/controllers/index.rb
@@ -9,13 +9,6 @@ get '/questions/:id' do
   erb :question
 end
 
-post '/questions/:question_id/answers/:answer_id/upvote' do
-  answer = Answer.find(params[:answer_id])
-  answer.scores.create(value: 1)
-  score = answer.scores.count
-  score.to_s
-end
-
 get '/question/new' do
   @user_id = session[:user_id]
   erb :'newquestion'
@@ -26,9 +19,46 @@ post '/questions' do
   redirect 'questions'
 end
 
+post '/questions/:question_id/upvote' do
+  if session[:user_id]
+    question = Question.find(params[:question_id])
+    question.scores.create(value: 1)
+    score = question.scores.count
+    score.to_s
+  else
+    return 'redirect'
+  end
+end
+
+post '/questions/:question_id/downvote' do
+  if session[:user_id]
+    question = Question.find(params[:question_id])
+    question.scores.last.destroy
+    score = question.scores.count
+    score.to_s
+  else
+    return 'redirect'
+  end
+end
+
+post '/questions/:question_id/answers/:answer_id/upvote' do
+  if session[:user_id]
+    answer = Answer.find(params[:answer_id])
+    answer.scores.create(value: 1)
+    score = answer.scores.count
+    score.to_s
+  else
+    return 'redirect'
+  end
+end
+
 post '/questions/:question_id/answers/:answer_id/downvote' do
-  answer = Answer.find(params[:answer_id])
-  answer.scores.last.destroy
-  score = answer.scores.count
-  score.to_s
+  if session[:user_id]
+    answer = Answer.find(params[:answer_id])
+    answer.scores.last.destroy
+    score = answer.scores.count
+    score.to_s
+  else
+    return 'redirect'
+  end
 end
